# openapi.yaml
openapi: 3.1.0
info:
  title: Books & Finance API
  version: 1.0.0
  description: |
    Personal finance and books management REST API.

    **Base URL:** `http://localhost:8000/api/v1`

    **Authentication:** Bearer JWT token. Obtain via `POST /auth/login`.
    Tokens expire in **7 days**.

    **Default credentials:** `BTC007` / `password123`
servers:
  - url: http://localhost:8000/api/v1
    description: Development server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  parameters:
    page:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
    limit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
...
    from:
      name: from
      in: query
      schema: { type: string, format: date }
    to:
      name: to
      in: query
      schema: { type: string, format: date }
security:
  - bearerAuth: []
paths:
...
  /income:
    get:
      summary: List income records
      tags: [Income]
      parameters:
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/limit'
        - $ref: '#/components/parameters/from'
        - $ref: '#/components/parameters/to'
      responses:
        '200': { description: Income list }
    post:
      summary: Add income
      tags: [Income]
      requestBody:
        required: true
        content:
          application/json:
            example:
              source: Freelance
              amount: 1500.00
              date: "2024-01-15"
              category: consulting
      responses:
        '201': { description: Income created }
...
  /public:
    get:
      summary: Public feed (no auth)
      security: []
      tags: [Public Feed]
      parameters:
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/limit'
        - name: type
          in: query
          schema: { type: string, enum: [update, tip, goal, achievement] }
      responses:
        '200':
          description: Public posts
          content:
            application/json:
              example:
                success: true
                data:
                  - id: 1
                    content: Just hit my savings goal!
                    type: achievement
                    likes: 14
                    created_at: "2024-01-15T12:00:00Z"
                    author: { username: BTC007 }
    post:
      summary: Create public post (auth required)
      tags: [Public Feed]
      requestBody:
        required: true
        content:
          application/json:
            example:
              content: Just saved my first ₹10,000!
              type: achievement
      responses:
        '201': { description: Post created }
  /public/{id}/like:
    patch:
      summary: Increment post likes (no auth, rate-limited)
      security: []
      tags: [Public Feed]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: Updated likes count
          content:
            application/json:
              example:
                success: true
                data: { id: 1, likes: 15 }
        '429': { description: Too many requests }